requirements.txt
streamlit
pandas
Pillow
python-dotenv
PyMuPDF
camelot-py
groq
google-generativeai
torch
transformers
opencv-python
regex